想法关联 (#16)
* Finish several tasks:
1. Create ideaLink page, connectCreateBox, connectIdeaDialog for add relationship for ideas.
2. Add ideaConnectManager for adding relationship and draw polylines.
3. Add clound function to create relationship in backend.
4. Add DomainManager to get domain information.

* feat(IdeaManager IdeaConnectManager DomainManager wedeaMap ideaLink): Draw polyline for links between ideas

The idea association is drawn, and the style of the line can be changed.
Idea drawing requires idea mapping, which is added to ideaMap. At the
same time, the construction of the mapping occurs after the setIdeas
event, so the timing of drawing lines is adjusted. WARNING: Now that
there are two maps, the incident has been dealt with initially, but
there may still be bugs.

BREAKING CHANGE: IdeaManager NEW ATTR, properly bugs of TWO MAP

* Fix a bug: when creating a new relationship, the map will not update.

* Fix a bug: Promise doesn't have finally part.

Co-authored-by: baitieyi <baitieyi@163.com>